Zadanie 12

+++ src/com/company/devices/Car.java
@@ -1,10 +1,12 @@
 package com.company.devices;
 import com.company.Human;
 import java.util.Objects;
+import java.util.List;
+import java.util.LinkedList;
 
 public abstract class Car extends Device
 {
-    //private Double value;
+    public LinkedList<Human> carOwners = new LinkedList<Human>();
 
     public Car(String producer, String model, Integer yearOfProduction, Double value)
     {
@@ -46,7 +48,7 @@ public abstract class Car extends Device
 
     public void sell(Human seller, Human buyer, Double price) throws Exception{
         if (buyer.freeSlots() == true) {
-            if (seller.isCar(this) == true) {
+            if (seller.isCar(this) == true && nowOwner(seller) == true) {
                 if (buyer.getCash() >= price) {
                     seller.deleteCar(this);
                     buyer.addCar(this);
@@ -60,7 +62,7 @@ public abstract class Car extends Device
                 }
             }
             else{
-                throw new Exception("Seller - no car!");
+                throw new Exception("Seller - no car! OR Seller - not last owner!");
                 //System.out.println("No car!");
             }
         }
@@ -70,4 +72,35 @@ public abstract class Car extends Device
         }
     }
     public abstract void refuel();
+    public void carOwner(Human carOwner) {
+        if (this.carOwners.contains(carOwner)) {
+            System.out.println("Was/Is owner");
+        }
+        else {
+            System.out.println("No owner!");
+        }
+    }
+    public boolean nowOwner(Human carOwner){
+        if (carOwners != null && !carOwners.isEmpty()){
+            if(carOwner == this.carOwners.getLast()){
+                return true;
+            }
+            return false;
+        }
+        return false;
+    }
+    public void sellCheck(Human seller, Human buyer) {
+        if (this.carOwners.indexOf(seller) == this.carOwners.indexOf(buyer) - 1) {
+            System.out.println(seller.firstName + " " + seller.lastName + " sold car: " + buyer.firstName + " " + buyer.lastName);
+        }
+        else {
+            System.out.println("No transation!");
+        }
+
+    }
+    public int transationNumber() {
+        int transactionNumber;
+        transactionNumber = this.carOwners.size() - 1;
+        return transactionNumber;
+    }
 }
